reserve also UNK token, more robust TokenTextEncoder
TokenTextEncoder is used in data_generators.wmt.ende_bpe_token_generator()
in the translate_ende_wmt_bpe32k problem, which uses an external BPE vocabulary
and training data pre-processed with these BPE subwords.
This may be useful when comparing with Nematus/NeuralMonkey etc. on exactly the same training data.
Unfortunately, the TokenTextEncoder implementation was not robust enough.
It failed if there was an unknown subword in the training or dev or test data,
e.g. because of unknown character.