trainer_lib: add problem hparams after parsing the overrides
t2t-trainer passes problem_name=None to create_hparams(), and
adds the problem hparams afterwards when creating the experiment.
t2t-decoder instead relies fully on create_hparams()
This inconsistency causes t2t-trainer and t2t-decoder to pass
different sets of model_hparams to Problem.hparams() (before and
after overrides respectively), which is problematic.

Always letting the problem see the final version of the model
hparams is more useful (in case the problem hparams depend on
the model), so make sure the problem is added after the overrides.