Fix KL regularization for Bayesian Layers.
Currently, computing derivatives of the sum of `layer.losses` w.r.t. to the
weight distribution variables yields `None` gradients in Eager mode.

This appears to be due to TF ops in the tfp.distribution `__init__` methods.
Since the regularization tensor is created at build time, those distribution
ops are tied to the context in which the distribution was created.  If we use a
`GradientTape`, and the layer gets built outside of the tape, the gradients
will always be equal to `None`.  If we build inside the `GradientTape`, say on
the first training epoch/step, the gradients will be computed correctly on that
iteration.  However, on the next iteration, in which a new tape is created, the
gradients will be equal to `None` again due to the TF ops that were created
inside the distribution's `__init__` method being tied to the first iteration's
tape.

This bug is particularly problematic because it won't appear to be broken if
the loss involves the usual likelihood term.  More specifically, there will be
gradients w.r.t. the variables, but they will only be from the likelihood term
and not from the KL term.  Additionally, the KL term will still produce a
value, so printing/summarizing that term will give the appearance that
everything is working.

This presents a fix that involves creating a callable for the "variable" when
creating the regularization tensor, with the condition that we must create a
new Edward RV within that callable.  A more permanent fix could involve making
changes to Edward RandomVariables or removing TF ops from the `__init__`
methods of TFP distributions.

PiperOrigin-RevId: 220898007