Changes for handling trajectory truncation.
EnvProblem - Add a method, `truncate` that calls truncate on the Trajectory.

env_problem_utils - Now if a trajectory reaches `max_timestep`, we don't call `reset` and let it continue. `reset` is only called when env returns done.

BatchTrajectory - Add a method, `truncate_trajectories` that puts the truncated trajectories in the completed bin, and adds new trajectories with the same starting observation and the truncated trajectory's last observation.

PiperOrigin-RevId: 249114001