[TRAX] Improve the efficiency of the multi-device _save_replicated function.
Profiling shows that save_replicated is responsible for triggering a large number of synchronous host to device transfers inside the call to .mean(0) on a ShardedDeviceArray.

Although we do also plan to improve the efficiency of the current code by improving JAX, we don't need to compute the mean at all since all replicas should be identical. Instead, use the value from the first replica.

PiperOrigin-RevId: 250893206