Add a flag --parallelize_envs that runs the envs in the EnvProblem in parallel.
If set to True, this uses the number of cpu cores, else runs the envs sequentially as in the current implementation.

Whether this actually speeds things up is indeterminate, rather for Atari this actually *slows* things down, probably because of the GIL.

Hopefully though for the `RPCEnv` and in processes where the step can take a long time, this should hopefully help a lot -- but this remains to be seen.

Let's check this in for now and try to fix it if the RPC code is slower.

PiperOrigin-RevId: 257704752