Add ability to stuff arbitrary data in a time-step and therefore a trajectory.
This helps for example to add log-probabilities of a selected action.

In addition to log-probabilities we also use it to stuff the value prediction
at each time-step as well.

Not only does this save us re-computing these things but is the right thing to
do in the face of randomness in the policy/value network (ex: dropout).

PiperOrigin-RevId: 257878599