Fix decoding in prepend mode (#1726)
* Create an integer problem_0_steps variable.

* Save inputs to the feature "partial_targets" when prepend_mode is not "none".

* Removed a second call to update_hparams_for_universal_transformer().

Fixes hyperparameter sets universal_transformer_big and universal_transformer_base_tpu.

* Fix a bug to make partial targets work for beam size > 1

The dimension of the multiplication of the partial targets was wrong:  (a, b, c, d) --> (a, b, c, d, a, b, c, d)
Correct multiplication needs to be: (a, b, c, d) --> (a, a, b, b, c, c, d, d)
This is because it is (batch_size * beam_size) instead of (beam_size * batch_size).

Basically, tf.tile needs to be replaced by tf.repeat which is introduced in tf 1.15. This is a workaround for tf 1.14.