Bump T2T version to 1.15.5
With this change Travis will turn green.

When we import compat.v1, we shouldn't do tf.compat.v1 anymore.

Also the following to unclog Travis:
six>=1.12.0 is required by tf-hub 0.8.0, one of the core dependencies seems to
be installing tf-hub, so pinning tf-hub to 0.7.0 (which is in the extras
section) doesn't seem to help.

Locally the previous setup.py installs fine, but only errors on Travis for some
reason.

PiperOrigin-RevId: 307203166