Deprecation notice for get_ende_bleu.sh
This script is harmful because it propagates a non-standard way to compute BLEU that is not reflective of the WMT 2014 task.  Entirely too many papers are submitted with BLEU scores computed in undocumented ways.  It's not even reasonable to allow people to run this script to compare against prior work, because most prior work does not document which script it used.  And there are multiple of these running around.  

https://www.aclweb.org/anthology/W18-6319/